tests reorganisation to have a main_tests hat setup rabbitmq and calls other tests
